refactor: add typed acceptance path alias and barrel imports
- replace inline PipelineContext acceptance path reference with explicit type alias

- switch changed files to path aliases via project barrels

- export needed barrel symbols to satisfy alias-internals lint

- keep lint and targeted acceptance tests green
