+++ src/execution/runner-completion.ts
@@ -6,23 +6,22 @@
  */
 
 import path from "node:path";
-import { groupStoriesByPackage } from "../acceptance/test-path";
-import type { NaxConfig } from "../config";
-import { loadConfigForWorkdir } from "../config/loader";
-import type { LoadedHooksConfig } from "../hooks";
-import { fireHook } from "../hooks";
-import { getSafeLogger } from "../logger";
-import type { StoryMetrics } from "../metrics";
-import type { PipelineEventEmitter } from "../pipeline/events";
-import type { AgentGetFn } from "../pipeline/types";
-import type { PluginRegistry } from "../plugins/registry";
-import { isComplete } from "../prd";
-import type { PRD } from "../prd";
-import { reviewOrchestrator } from "../review/orchestrator";
-import type { DispatchContext } from "../runtime/dispatch-context";
-import type { ISessionManager } from "../session";
-import { errorMessage } from "../utils/errors";
-import { autoCommitIfDirty } from "../utils/git";
+import { groupStoriesByPackage } from "@/acceptance";
+import { type NaxConfig, loadConfigForWorkdir } from "@/config";
+import type { LoadedHooksConfig } from "@/hooks";
+import { fireHook } from "@/hooks";
+import { getSafeLogger } from "@/logger";
+import type { StoryMetrics } from "@/metrics";
+import type { PipelineEventEmitter } from "@/pipeline/events";
+import type { AgentGetFn } from "@/pipeline/types";
+import type { PluginRegistry } from "@/plugins/registry";
+import { isComplete } from "@/prd";
+import type { PRD } from "@/prd";
+import { reviewOrchestrator } from "@/review";
+import type { DispatchContext } from "@/runtime/dispatch-context";
+import type { ISessionManager } from "@/session";
+import { errorMessage } from "@/utils/errors";
+import { autoCommitIfDirty } from "@/utils/git";
 import { stopHeartbeat, writeExitSummary } from "./crash-recovery";
 import type { AcceptanceLoopContext, AcceptanceLoopResult } from "./lifecycle/acceptance-loop";
 import type { RunCompletionOptions, RunCompletionResult } from "./lifecycle/run-completion";
