+++ src/review/index.ts
@@ -8,5 +8,6 @@ export * from "./adversarial";
 export * from "./categorization";
 export * from "./diff-utils";
 export * from "./finding-projection";
+export * from "./orchestrator";
 export * from "./types";
 export * from "./runner";
